Switch execution order of scrub_items() and prune_news()
Rationale was that cache scrub would reduce number of items to iterate
through later. However, if expiration hasn't been updated in "long"
time eg. due to error or low traffic(HTTP Staus Code 304), then the
whole cache gets invalidated which leads to excess flood.

One solution is to increase cache expiration(in other words one day is
really low).
Another solution seems to be to switch order of execution, because
prune_news() updates time of expiration. Therefore, it should prevent
scrubbing of news which are still present in RSS feed.
And it's also possible there is a bug somewhere in the code.